Updated Retrofit, Fixed Mulitple Issue With View Binding
- Binding view model fixes
- Update retrofit v2.3.0 -> v2.4.0

+++ app/src/main/java/com/mxt/anitrend/base/custom/view/text/RatingTextView.java
@@ -67,7 +67,7 @@ public class RatingTextView extends LinearLayout implements CustomView {
     }
 
     @BindingAdapter("rating")
-    public static void setAverageRating(RatingTextView view, double averageScore) {
+    public static void setAverageRating(RatingTextView view, int averageScore) {
         float rating = (float)averageScore * MAX / 100;
         view.setRating(rating);
     }
