+++ app/src/main/java/com/mxt/anitrend/base/custom/view/widget/CustomRatingBar.java
@@ -49,7 +49,7 @@ public class CustomRatingBar extends AppCompatRatingBar implements CustomView {
     }
 
     @BindingAdapter("averageScore")
-    public static void setAverageScore(CustomRatingBar view, double averageScore) {
+    public static void setAverageScore(CustomRatingBar view, int averageScore) {
         float rating = (float)averageScore * view.getMax() / 100;
         view.setRating(rating);
     }
